+++ src/test/zeromq/middleware_regex/test.cpp
@@ -26,7 +26,7 @@
 #include <deque>
 
 #include "goby/middleware/marshalling/protobuf.h"
-#include "goby/middleware/transport.h"
+#include "goby/middleware/transport/interthread.h"
 #include "goby/util/debug_logger.h"
 #include "goby/zeromq/transport/interprocess.h"
