+++ src/test/zeromq/middleware_speed/test.cpp
@@ -28,7 +28,9 @@
 #include <boost/units/io.hpp>
 
 #include "goby/middleware/marshalling/protobuf.h"
-#include "goby/middleware/transport.h"
+#include "goby/middleware/transport/interthread.h"
+#include "goby/time/system_clock.h"
+#include "goby/time/types.h"
 #include "goby/util/debug_logger.h"
 #include "goby/zeromq/transport/interprocess.h"
