+++ src/test/zeromq/zeromq_and_intervehicle/test.cpp
@@ -27,7 +27,7 @@
 
 #include "goby/middleware/marshalling/dccl.h"
 #include "goby/middleware/marshalling/protobuf.h"
-#include "goby/middleware/transport.h"
+#include "goby/middleware/transport/intervehicle.h"
 #include "goby/util/debug_logger.h"
 #include "goby/zeromq/transport/interprocess.h"
